Web_o2b2_main/O2B2WEB_DB-main/O2B2WEB-main/cradle-to-grave/job/requirements.txt
streamlit
python-dotenv
pymongo
langchain-core
langchain-openai
Pillow
langchain
langchain-community